Update contact info, and footer, links
Now contact info shall be displayed at the bottom of the site, removed the ugly line under links, as well changed the color of links, that means that they will be displayed as orange at the start, if the mouse hovers over them, they shall become a light blue, and when clicked a darker blue. Background img hopefully fixed, won't have issues resizing to other resolutions.